// src/utils/anime-easing-enum.ts
export enum AnimeEasingNames {
    easeInSine = 'easeInSine',
    easeOutSine = 'easeOutSine',
    easeInOutSine = 'easeInOutSine',
    easeInCirc = 'easeInCirc',
    easeOutCirc = 'easeOutCirc',
    easeInOutCirc = 'easeInOutCirc',
    easeInElastic = 'easeInElastic',
    easeOutElastic = 'easeOutElastic',
    easeInOutElastic = 'easeInOutElastic',
    easeInBack = 'easeInBack',
    easeOutBack = 'easeOutBack',
    easeInOutBack = 'easeInOutBack',
    easeInBounce = 'easeInBounce',
    easeOutBounce = 'easeOutBounce',
    easeInOutBounce = 'easeInOutBounce',
    easeInQuad = 'easeInQuad',
    easeOutQuad = 'easeOutQuad',
    easeInOutQuad = 'easeInOutQuad',
    easeInCubic = 'easeInCubic',
    easeOutCubic = 'easeOutCubic',
    easeInOutCubic = 'easeInOutCubic',
    easeInQuart = 'easeInQuart',
    easeOutQuart = 'easeOutQuart',
    easeInOutQuart = 'easeInOutQuart',
    easeInQuint = 'easeInQuint',
    easeOutQuint = 'easeOutQuint',
    easeInOutQuint = 'easeInOutQuint',
    easeInExpo = 'easeInExpo',
    easeOutExpo = 'easeOutExpo',
    easeInOutExpo = 'easeInOutExpo',
    linear = 'linear'
}